backend/requirements.txt
Django==5.0.6
channels==4.1.0
djangorestframework==3.15.2
selenium==4.21.0
pytest==8.2.2